+++ CHANGES.md
@@ -56,6 +56,8 @@
         - `"libwebrtc": "Shiguredo-Build M79@{#5} (79.5.4 b484ec00)"`
         - `"ayameClient": "WebRTC Native Client Momo 2020.1 (f6b69e77)"`
     - @melpon
+- [CHANGE] momo --help の英語化
+    - @shino msnoigrs
 - [CHANGE] <package>.edit の機能とドキュメントを削除
     - @melpon
 - [FIX] --no-video を指定しているにもかかわらずカメラを一瞬だけ掴むのを修正する