+++ CHANGES.md
@@ -11,6 +11,9 @@
 
 ## develop
 
+- [UPDATE] libwebrtc を M80.3987@{#2} に上げる
+    - libwebrtc のハッシュは fba51dc69b97f6f170d9c325a38e05ddd69c8b28
+    - @melpon
 - [UPDATE] Momo 2020.1 にバージョンを上げる
     - バージョン番号を <リリース年>.<その年のリリース回数> に変更
     - @voluntas